Potential solution into finding m1 and s1
So I looked around and apparently, b1 is E F G H I J, which is left-to-right above, and h1 is A C G M R T which is top-to-bottom from left. Naturally, I decided to make m1 and s1 similar to b1 and h1, but m1 is like b1 except below and s1 is like h1 but from the right side.