Use dedicated structure for Intra prediction samples
Previously, it was just a simple span - for 4x4, 8x8 and 16x16, it was 5x5, 9x9 and 17x17 to make room for left, top and left-top pixels.

This was a bit hard to maintain and debug, so to make everything more clear, we now use a new structure, IntraPredictionSamples, over a simple span of integers.